pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag_agent"
version = "0.1.0"
description = "Modular, port–adapter RAG agent with Gradio UI"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"

# Core dependencies (interface + LangGraph)
dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "langchain-core>=0.3",
  "langgraph>=0.6",
]

[project.optional-dependencies]
llm = [
  "openai>=1.13",
  "langchain-ollama>=0.3",
  "ollama>=0.1.10",
  "langchain_community>=0.3.1"
]
vector = [
  "langchain-chroma>=0.2.5",
  "chromadb>=1.0.15",
]
ui = [
  "gradio>=5.0",
]
loader = [
  "unstructured[docx,html]>=0.18.11",  # for PDF/HTML partitioning
  "python-docx>=1.2.0",               # for DOCX parsing
  "pdf2image>=1.17.0",                # OCR fallback
  "pytesseract>=0.3.10",              # OCR engine
  "beautifulsoup4>=4.13.0",           # optional fallback for HTML
]
reranker = [
  "transformers>=4.30",
  "torch>=2.0"
]
dev = [
  "pytest>=8",
  "pytest-asyncio>=1.1.0",
  "ruff>=0.3",
  "mypy>=1.8",
]
all = [
  "openai>=1.13",
  "langchain-ollama>=0.3",
  "ollama>=0.1.10",
  "langchain_community>=0.3.1",
  "langchain-chroma>=0.2.5",
  "chromadb>=1.0.15",
  "gradio>=5.0",
  "unstructured[docx,html]>=0.18.11",
  "python-docx>=1.2.0",
  "pdf2image>=1.17.0",
  "pytesseract>=0.3.10",
  "beautifulsoup4>=4.13.0",
  "transformers>=4.30",
  "torch>=2.0",
  "python-dotenv>=1.0"
]
windows = ["python-magic-bin>=0.4.14"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["rag_agent*"]